Add two numbers solution

+++ CODEFORCES SHEET 1/G.SummationFrom1toN.cpp
@@ -15,7 +15,7 @@
 using namespace std;
 
 int main() {
-    int N;
+    long long N;
     cin >> N;
 
     cout << (N * (N + 1)) / 2 << endl;
